finished

+++ internal/models/area.go
@@ -1,10 +1,10 @@
 package models
 
 type AreaCheck struct {
-	X        int    `json:"x"`
-	Y        int    `json:"y"`
-	Radius   int    `json:"radius"`
-	Result   bool   `json:"result"`
-	SendTime string `json:"send_time"`
-	Author   string `json:"author"`
+	X        float64 `json:"x"`
+	Y        float64 `json:"y"`
+	Radius   float64     `json:"radius"`
+	Result   bool    `json:"result"`
+	SendTime string  `json:"send_time"`
+	Author   string  `json:"author"`
 }
